fixed settings screen, added new headerRight component
